fix: Modal overlay crop issue (#31375)
