+++ app/client/src/layoutSystems/common/modalOverlay/ModalOverlayLayer.tsx
@@ -11,6 +11,7 @@ import { selectWidgetInitAction } from "actions/widgetSelectionActions";
 import { SelectionRequestType } from "sagas/WidgetSelectUtils";
 import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
 import { useMaxModalWidth } from "widgets/ModalWidget/component/useModalWidth";
+import { useAppViewerSidebarProperties } from "utils/hooks/useAppViewerSidebarProperties";
 const Container = styled.div<{
   width?: number;
   height?: number;
@@ -24,6 +25,7 @@ const Container = styled.div<{
   isEditMode?: boolean;
   headerHeight?: number;
   smallHeaderHeight?: string;
+  leftSidebarWidth?: string;
 }>`
   &&& {
     .${Classes.OVERLAY} {
@@ -34,9 +36,13 @@ const Container = styled.div<{
       top: 0;
       right: 0;
       bottom: 0;
+      left: ${(props) => props.leftSidebarWidth || 0}px;
       height: 100%;
       z-index: ${(props) => props.zIndex};
-      width: 100%;
+      width: ${(props) =>
+        props.leftSidebarWidth !== "0"
+          ? `calc(100% - ${props.leftSidebarWidth})`
+          : "100%"};
       display: flex;
       justify-content: center;
       align-items: center;
@@ -123,6 +129,7 @@ export function ModalOverlayLayer(props: BaseWidgetProps) {
   };
 
   const maxModalWidth = useMaxModalWidth();
+  const { hasSidebarPinned, sidebarWidth } = useAppViewerSidebarProperties();
 
   return (
     <ComponentContainerWrapper isEditMode={props.isEditMode}>
@@ -142,6 +149,7 @@ export function ModalOverlayLayer(props: BaseWidgetProps) {
           height={props.height}
           isEditMode={props.isEditMode}
           left={props.left}
+          leftSidebarWidth={hasSidebarPinned ? sidebarWidth.toString() : "0"}
           maxWidth={maxModalWidth}
           minSize={props.minSize}
           right={props.bottom}
