+++ app/client/src/pages/AppViewer/AppViewerPageContainer.tsx
@@ -9,11 +9,7 @@ import { theme } from "constants/DefaultTheme";
 import { Icon, NonIdealState, Spinner } from "@blueprintjs/core";
 import Centered from "components/designSystems/appsmith/CenteredWrapper";
 import AppPage from "./AppPage";
-import {
-  getCanvasWidth,
-  getCurrentPageName,
-  getIsAutoLayout,
-} from "selectors/editorSelectors";
+import { getCanvasWidth, getCurrentPageName } from "selectors/editorSelectors";
 import RequestConfirmationModal from "pages/Editor/RequestConfirmationModal";
 import { getCurrentApplication } from "@appsmith/selectors/applicationSelectors";
 import {
@@ -24,14 +20,13 @@ import { builderURL } from "@appsmith/RouteBuilder";
 import { getCanvasWidgetsStructure } from "@appsmith/selectors/entitiesSelector";
 import equal from "fast-deep-equal/es6";
 
-const Section = styled.section<{ $contain: string }>`
+const Section = styled.section`
   height: 100%;
   width: 100%;
   margin: 0 auto;
   position: relative;
   overflow-x: auto;
   overflow-y: auto;
-  contain: ${({ $contain }) => $contain};
 `;
 
 type AppViewerPageContainerProps = RouteComponentProps<AppViewerRouteParams>;
@@ -42,7 +37,6 @@ function AppViewerPageContainer(props: AppViewerPageContainerProps) {
   const canvasWidth = useSelector(getCanvasWidth);
   const isFetchingPage = useSelector(getIsFetchingPage);
   const currentApplication = useSelector(getCurrentApplication);
-  const isAutoLayout = useSelector(getIsAutoLayout);
   const { match } = props;
 
   // get appsmith editr link
@@ -97,7 +91,7 @@ function AppViewerPageContainer(props: AppViewerPageContainerProps) {
     return pageNotFound;
 
   return (
-    <Section $contain={isAutoLayout ? "content" : "strict"}>
+    <Section>
       <AppPage
         appName={currentApplication?.name}
         canvasWidth={canvasWidth}
